Add ability to test a single file (#1)
You can mark a test with `.only()` in [Ava](https://github.com/avajs/ava) but it doesn't mean only that test will run. The whole suite will still run but when it comes to that test file just the test marked as `.only()` will run in it. In our opinion, that makes it pretty pointless.

So much so the amazing @StuAA78 looked into whether he could solve it. We are VSCode users and love a [custom task](https://code.visualstudio.com/docs/editor/tasks). @StuAA78 found you can get the current file using `$relativeFile`.

```json
  {
    "label": "✅ Test only",
    "detail": "Run just this file's unit tests",
    "type": "shell",
    "command": "npm test ${relativeFile}",
    "group": "test",
    "presentation": {
      "echo": true,
      "reveal": "always",
      "focus": true,
      "panel": "shared",
      "showReuseMessage": false,
      "clear": true
    }
  }
```

So, if you have a test file open and use this custom VSCode task, it will tell **Ava** to just run that one. Now `.only()` becomes useful again to just focus on a single test in a single file. But he didn't stop there. What happens if you don't have `*.test.js` file open in VSCode?

@StuAA78 came up with a bash script that will attempt to identify the matching test file and pass that to **Ava**.

```bash

if [ $# -eq 0 ]; then
  echo "Usage: $0 <filename>"
  exit 1
fi

filename="$1"

if [[ $filename != *".test."* ]]; then
  # Extract the path and file parts
  path="$(dirname "$filename")"
  file="$(basename "$filename")"

  # Replace ".js" with ".test.js" in the file part
  test_file="${file/.js/.test.js}"

  # Build the new path, replacing the first part with "test"
  path_parts=($path)
  new_path="test"
  for ((i = 1; i < ${#path_parts[@]}; i++)); do
    new_path="$new_path/${path_parts[$i]}"
  done

  # Construct the new filename
  filename="$new_path/$test_file"
fi

npm test -- $filename
```

Run that instead of `npm test ${relativeFile}` and that scenario is dealt with. Or so we hoped. It didn't seem to work in our Dockerised local environment. The part of the script that breaks up the parts never got triggered which meant everything became `test/*.test.js`. It also requires you to provide a file and to call something other than `npm test`.

Based on this inspiring solution we're making this change. Instead of bash, we're using Node. We always know that will be available as opposed to bash, which is dependent on the OS and setup. It also makes the file name manipulation easier. With these changes, you always call `npm test` to be consistent. It is just whether you add an extra arg that running just one file kicks in.

Who knows, this might inspire someone else to make an even better solution to this problem!